Update for OSX to include data files/
Very crude update to PluginLocalization.cmake to get the data files in
the package.
A problem is that under OSX the browser does not let you go into
OpenCPN.app where the polar data files are.
So you have to select OpenCPN.
Plugin will complain that there is no data file.
You have to manually add the / sign.
Plugin will complain again and then it is possible to browse further.