Fix crash when editing polars caused by spurious paint events.
For some reason, some spurious paint events are generated by the BoatDialog
at the same time the user tries to edit or modify the polar. The code in the
paint event handler accesses the polar data in an unprotected way, and reads
uninitialized data at some point.

This commit fixes this issue by blocking the paint event handlers while the
EditPolarDialog is shown.