Add a proper RPATH (#1977).
Add a $ORIGIN-based rpath to the plugin so that the ld.so dynamic
linker can locate plugin libraries also on flatpak, where the
LD_LIBRARY_PATH mechanism which is used on native Linux cannot be
used (the flatpak sandbox reserves this for internal use, since it
might introduce dependencies outside the sandbox).

Currently, we have no dynamic libraries used. However, using an
rpath is a good habit which doesn't change anything if not
required, while solving an otherwise tricky problem.